Refactor: reorganize repo package into sub-packages.
Moved `autocomplete`, `history`, and `macro` into dedicated sub-packages for better modularity and clarity. Updated imports and references across the codebase to reflect the new structure.
