+++ pkg/core/command/factory_test.go
@@ -7,12 +7,12 @@ import (
 	"time"
 
 	"github.com/ksysoev/wsget/pkg/core"
-	"github.com/ksysoev/wsget/pkg/repo"
+	"github.com/ksysoev/wsget/pkg/repo/macro"
 )
 
 func TestFactory_Create(t *testing.T) {
 	tests := []struct {
-		macro   *repo.Macro
+		macro   *macro.Macro
 		want    core.Executer
 		name    string
 		raw     string
@@ -91,7 +91,7 @@ func TestFactory_Create(t *testing.T) {
 		{
 			name:    "macro command",
 			raw:     "macro",
-			macro:   &repo.Macro{},
+			macro:   &macro.Macro{},
 			want:    nil,
 			wantErr: true,
 		},
