+++ pkg/repo/history/autocomplete.go
@@ -1,4 +1,4 @@
-package repo
+package history
 
 import (
 	"sort"
