+++ pkg/repo/history/autocomplete_test.go
@@ -1,4 +1,4 @@
-package repo
+package history
 
 import (
 	"testing"
