+++ pkg/repo/history/history.go
@@ -1,4 +1,4 @@
-package repo
+package history
 
 import (
 	"bufio"
