+++ pkg/repo/macro/macro.go
@@ -1,4 +1,4 @@
-package repo
+package macro
 
 import (
 	"bytes"
