+++ pkg/repo/macro/macro_test.go
@@ -1,4 +1,4 @@
-package repo
+package macro
 
 import (
 	"os"
